requirements.txt
environ
certifi==2019.11.28
chardet==3.0.4
cheroot==8.3.0
CherryPy==18.5.0
dj-database-url==0.5.0
Django==2.2.5
django-crispy-forms==1.7.2
django-extensions==2.2.1
django-paypal==1.0.0
gunicorn==20.0.4
idna==2.8
jaraco.classes==3.1.0
jaraco.collections==3.0.0
jaraco.functools==3.0.0
jaraco.text==3.2.0
more-itertools==8.2.0
Pillow==6.1.0
portend==2.6
psycopg2==2.8.3
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
tempora==3.0.0
urllib3==1.25.8
Webhook-Listener==0.0.1
whitenoise==5.0.1
zc.lockfile==2.0
